refactor: split out messageCreate handler
