+++ src/command/commands/games/zip/zip-remove-submission.ts
@@ -1,7 +1,7 @@
 import { ChatInputCommandInteraction, SlashCommandBuilder, Snowflake } from "discord.js";
 import { Command } from "../../../command";
 
-import { ZipGameDatabase } from "../../../../games/zipgamedb";
+import { ZipGameDatabase } from "../../../../fun/zipgamedb";
 
 export const command: Command = {
 	data: new SlashCommandBuilder()
