+++ src/command/commands/games/zombbblob/end-zombbblob.ts
@@ -1,7 +1,7 @@
 import { ChatInputCommandInteraction, SlashCommandBuilder } from "discord.js";
 import { Command } from "../../../command";
 
-import { WordsDatabase } from "../../../../games/zombbblobdb";
+import { WordsDatabase } from "../../../../fun/zombbblobdb";
 
 export const command: Command = {
 	data: new SlashCommandBuilder()
