+++ src/command/commands/games/zombbblob/reroll.ts
@@ -1,7 +1,7 @@
 import { ChatInputCommandInteraction, SlashCommandBuilder } from "discord.js";
 import { Command } from "../../../command";
 
-import { WordsDatabase } from "../../../../games/zombbblobdb";
+import { WordsDatabase } from "../../../../fun/zombbblobdb";
 import { ConfigHandler } from '../../../../config/config';
 import { maintainersPingString } from "../../../../utils";
 const { Roles, Channels } = ConfigHandler.getInstance().getConfig();
