+++ src/fun/zipgame.ts
