+++ src/fun/zombiegame.ts
